Create a R project for the brief report and started codeing
The repository now includes a R project. A data file with life tables and an .R file with the code.

We start with a time homogeneous part and show how to create M and A. Also how a matrix for the cost and utilities is created and how to apply this.